updatde

+++ src/ebike-panel/ebike-panel-sharedComponent/all-bikes-card/index.tsx
@@ -1,9 +1,9 @@
 'use client'
 import { ChangeApprove, ChangeFeatured, DeleteBlogById, DeleteNewBikeById, DeleteUsedBikeById, getAllBlog, getAllNewBike, getCustomBikeAd } from "@/ebike-panel/ebike-panel-Function/globalfunction";
 import { getBrandFromId, getCityFromId } from "@/ebikeWeb/functions/globalFuntions";
+import { add3Dots, priceWithCommas } from "@/genericFunctions/geneFunc";
 import { BrandArr, CityArr } from "@/ebikeWeb/constants/globalData";
 import Loader from "@/ebikeWeb/sharedComponents/loader/loader";
-import { add3Dots, priceWithCommas } from "@/genericFunctions/geneFunc";
 import { Navigation, FreeMode } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Grid, Pagination } from "@mui/material";
@@ -14,8 +14,8 @@ import styles from './index.module.scss';
 import '../../../app/globals.scss';
 import 'swiper/css/navigation';
 import 'swiper/css';
-import { FaLaptopHouse } from "react-icons/fa";
 
+/////////////////////////////////////////////////////// USED BIKE CARD
 const Used_bike_card = () => {
 
     const [displayedBikes, setDisplayedBikes] = useState([]);
@@ -95,7 +95,6 @@ const Used_bike_card = () => {
         }
     };
 
-
     const handleFeatureToggle = async (id: any, currentStatus: boolean) => {
         const confirmDelete = window.confirm("Are you sure to change Featured?");
         if (!confirmDelete) return;
@@ -153,9 +152,12 @@ const Used_bike_card = () => {
                 } else {
                     setAllBikeForFilter([]);
                 }
-
                 setCurrentPage(_page);
                 setAllBikeArr(res.data);
+                setIsLoading(false);
+                setTimeout(() => {
+                    window.scrollTo(0, 0)
+                }, 500);
             } else {
                 setCurrentPage(1);
                 setAllBikeArr([]);
@@ -165,8 +167,6 @@ const Used_bike_card = () => {
             }
         } catch (error) {
             console.error("Error fetching bikes:", error);
-        } finally {
-            setIsLoading(false);
         }
     };
 
@@ -181,7 +181,7 @@ const Used_bike_card = () => {
                                 <div className={styles.main_box_card} key={i}>
                                     <div className={styles.card_container_box}>
                                         <div className={styles.card_header}>
-                                            <h3 className={styles.heading}>{add3Dots(e?.title , 50) || 'No Title'}</h3>
+                                            <h3 className={styles.heading}>{add3Dots(e?.title, 50) || 'No Title'}</h3>
                                             <span className={`${styles.featured_badge} ${e?.isFeatured ? styles.featured : ''}`}>
                                                 IsFeatured: {e?.isFeatured ? 'True' : 'False'}
                                             </span>
@@ -310,10 +310,11 @@ const Used_bike_card = () => {
     );
 };
 
+/////////////////////////////////////////////////////// NEW BIKE CARD
 const New_bike_card = () => {
-    const [AllNewBikeForFilter, setAllNewBikeForFilter] = useState([]);
-    const [filteredBikes, setFilteredBikes] = useState([]);
-    const [displayedBikes, setDisplayedBikes] = useState([]);
+    const [AllNewBikeForFilter, setAllNewBikeForFilter] = useState<any>([]);
+    const [filteredBikes, setFilteredBikes] = useState<any>([]);
+    const [displayedBikes, setDisplayedBikes] = useState<any>([]);
     const [searchTerm, setSearchTerm] = useState("");
     const [currentPage, setCurrentPage] = useState(1);
     const [totalPage, setTotalPage] = useState<any>(null);
@@ -356,6 +357,7 @@ const New_bike_card = () => {
                 setCurrentPage(1);
                 setTotalPage(0);
             }
+
         } catch (error) {
             console.error("Error fetching new bikes:", error);
         }
@@ -381,11 +383,14 @@ const New_bike_card = () => {
     };
 
     const handleDelete = async (id: any) => {
+        const isConfirmed = window.confirm("Are you sure you want to delete this bike?");
+
+        if (!isConfirmed) return;
         const res = await DeleteNewBikeById(id);
-        if(res && res.deleted){
+        if (res && res.deleted) {
             fetchAllNewBike(currentPage);
         }
-        else{
+        else {
             alert('Something is Wrong!')
         }
         console.log("Deleted:", res);
@@ -405,80 +410,86 @@ const New_bike_card = () => {
 
             {!IsLoading ? (
                 <div>
-                    {displayedBikes.map((e: any, i: any) => (
+                    {displayedBikes.length > 0 ? (
+                        <>
+                            {displayedBikes.map((e: any, i: any) => (
+                                <div className={styles.main_box_card} key={i}>
+                                    <div className={styles.card_container_box}>
+                                        <div className={styles.card_header}>
+                                            <h3 className={styles.heading}>{add3Dots(e?.title, 45) || 'No Title'}</h3>
+                                            <span className={styles.ad_id}>Ad ID: {e?.id}</span>
+                                        </div>
 
-                        <div className={styles.main_box_card} key={i}>
-                            <div className={styles.card_container_box}>
-                                <div className={styles.card_header}>
-                                    <h3 className={styles.heading}>{e?.title}</h3>
-                                    <span className={styles.ad_id}>Ad ID: {e?.id}</span>
-                                </div>
-
-                                <div className={styles.card_content}>
-                                    <div className={styles.cardimage_box}>
-                                        <Swiper
-                                            spaceBetween={50}
-                                            slidesPerView={1}
-                                            onSlideChange={() => console.log('slide change')}
-                                            onSwiper={(swiper) => console.log(swiper)}
-                                            modules={[Navigation, FreeMode]}
-                                            navigation={true}
-                                            initialSlide={0}
-                                            loop={true}
-                                            className={styles.image}
-                                        >
-                                            {
-                                                e?.images && e?.images.length > 0 ?
-                                                    e.images.map((imgUrl: any, ind: any) => {
-                                                        return (
-                                                            <SwiperSlide key={imgUrl} className={styles.image} >
-                                                                <img src={imgUrl} alt={e?.title} className={styles.image} />
+                                        <div className={styles.card_content}>
+                                            <div className={styles.cardimage_box}>
+                                                <Swiper
+                                                    spaceBetween={50}
+                                                    slidesPerView={1}
+                                                    onSlideChange={() => console.log('slide change')}
+                                                    onSwiper={(swiper) => console.log(swiper)}
+                                                    modules={[Navigation, FreeMode]}
+                                                    navigation={true}
+                                                    initialSlide={0}
+                                                    loop={true}
+                                                    className={styles.image}
+                                                >
+                                                    {
+                                                        e?.images && e?.images.length > 0 ?
+                                                            e.images.map((imgUrl: any, ind: any) => {
+                                                                return (
+                                                                    <SwiperSlide key={imgUrl} className={styles.image} >
+                                                                        <img src={imgUrl} alt={e?.title} className={styles.image} />
+                                                                    </SwiperSlide>
+                                                                )
+                                                            }) :
+                                                            <SwiperSlide key=''>
+                                                                <img src='https://res.cloudinary.com/dtroqldun/image/upload/c_scale,f_auto,h_200,q_auto,w_auto,dpr_auto/v1549082792/ebike-graphics/placeholders/used_bike_default_pic.png' alt={e?.title} className={styles.slider_img} />
                                                             </SwiperSlide>
-                                                        )
-                                                    }) :
-                                                    <SwiperSlide key=''>
-                                                        <img src='https://res.cloudinary.com/dtroqldun/image/upload/c_scale,f_auto,h_200,q_auto,w_auto,dpr_auto/v1549082792/ebike-graphics/placeholders/used_bike_default_pic.png' alt={e?.title} className={styles.slider_img} />
-                                                    </SwiperSlide>
-                                            }
-                                        </Swiper>
-                                    </div>
+                                                    }
+                                                </Swiper>
+                                            </div>
 
-                                    <div className={styles.card_detail}>
-                                        <div className={styles.detail_row}>
-                                            <span className={styles.detail_label}>ID:</span>
-                                            <span>{e?.id}</span>
+                                            <div className={styles.card_detail}>
+                                                <div className={styles.detail_row}>
+                                                    <span className={styles.detail_label}>ID:</span>
+                                                    <span>{e?.id}</span>
+                                                </div>
+
+                                                <div className={styles.detail_row}>
+                                                    <span className={styles.detail_label}>Brand:</span>
+                                                    <span>{GetName("brand", e?.brandId)}</span>
+                                                </div>
+
+                                                <div className={styles.detail_row}>
+                                                    <span className={styles.detail_label}>Price:</span>
+                                                    <span className={styles.price}>
+                                                        {e?.price ? priceWithCommas(e.price) : '0'}
+                                                    </span>
+                                                </div>
+
+                                                <div className={styles.detail_row}>
+                                                    <p style={{ margin: "10px 0px", padding: "0px", color: "black", fontSize: "14px" }} className={styles.desc} dangerouslySetInnerHTML={{ __html: e?.description }}></p>
+                                                </div>
+                                            </div>
                                         </div>
 
-                                        <div className={styles.detail_row}>
-                                            <span className={styles.detail_label}>Brand:</span>
-                                            <span>{GetName("brand", e?.brandId)}</span>
-                                        </div>
-
-                                        <div className={styles.detail_row}>
-                                            <span className={styles.detail_label}>Price:</span>
-                                            <span className={styles.price}>
-                                                {e?.price ? priceWithCommas(e.price) : '0'}
-                                            </span>
-                                        </div>
-
-                                        <div className={styles.detail_row}>
-                                            <p style={{ margin: "10px 0px", padding: "0px", color: "black", fontSize: "14px" }} className={styles.desc} dangerouslySetInnerHTML={{ __html: e?.description }}></p>
+                                        <div className={styles.card_actions}>
+                                            <button className={`${styles.action_btn} ${styles.edit_btn}`} onClick={() => handleEdit(e?.id)}>
+                                                <a href={`/ebike-panel/dashboard/edit-new-bike/${e?.id}`} style={{ textDecoration: 'none', color: "white" }}>
+                                                    Edit
+                                                </a></button>
+                                            <button className={`${styles.action_btn} ${styles.delete_btn}`} onClick={() => handleDelete(e?.id)}>
+                                                Delete
+                                            </button>
                                         </div>
                                     </div>
                                 </div>
-
-                                <div className={styles.card_actions}>
-                                    <button className={`${styles.action_btn} ${styles.edit_btn}`} onClick={() => handleEdit(e?.id)}>
-                                        <a href={`/ebike-panel/dashboard/edit-new-bike/${e?.id}`} style={{ textDecoration: 'none', color: "white" }}>
-                                            Edit
-                                        </a></button>
-                                    <button className={`${styles.action_btn} ${styles.delete_btn}`} onClick={() => handleDelete(e?.id)}>
-                                        Delete
-                                    </button>
-                                </div>
-                            </div>
+                            ))}
+                        </>) : (
+                        <div className={styles.no_results}>
+                            <p>No bikes found matching your search criteria.</p>
                         </div>
-                    ))}
+                    )}
                     <div className={styles.pagination}>
                         {filteredBikes?.length > 0 && (
                             <div className={styles.used_bike_list_pagination}>
@@ -503,6 +514,7 @@ const New_bike_card = () => {
     );
 };
 
+/////////////////////////////////////////////////////// BLOG CARD
 const Blog_Card = () => {
     const [AllBlogFilter, setAllBlogFilter] = useState([]);
     const [filteredBlog, setFilteredBlog] = useState([]);
@@ -565,6 +577,8 @@ const Blog_Card = () => {
     };
 
     const handleDelete = async (id: any) => {
+        const isConfirm = window.confirm('Are you sure to delete this blog?')
+        if (!isConfirm) return;
         const res = await DeleteBlogById(id);
         if (res && res.info == 'Blog has been deleted') {
             fetchAllBlog(currentPage);
@@ -593,7 +607,7 @@ const Blog_Card = () => {
                         <div className={styles.main_box_card} key={i}>
                             <div className={styles.card_container_box}>
                                 <div className={styles.card_header}>
-                                    <h3 className={styles.heading}>{e?.blogTitle}</h3>
+                                    <h3 className={styles.heading}>{add3Dots(e?.blogTitle, 50) || 'No Title'}</h3>
                                     <span className={styles.ad_id}>Ad ID: {e?.id}</span>
                                 </div>
